+++ x-pack/plugins/enterprise_search/server/routes/app_search/engines.ts
@@ -95,51 +95,4 @@ export function registerEnginesRoutes({
       path: '/as/engines/:name/overview_metrics',
     })
   );
-  router.get(
-    {
-      path: '/api/app_search/engines/{name}/source_engines',
-      validate: {
-        params: schema.object({
-          name: schema.string(),
-        }),
-        query: schema.object({
-          'page[current]': schema.number(),
-          'page[size]': schema.number(),
-        }),
-      },
-    },
-    enterpriseSearchRequestHandler.createRequest({
-      path: '/as/engines/:name/source_engines',
-    })
-  );
-  router.post(
-    {
-      path: '/api/app_search/engines/{name}/source_engines/bulk_create',
-      validate: {
-        params: schema.object({
-          name: schema.string(),
-        }),
-        body: schema.object({
-          source_engine_slugs: schema.arrayOf(schema.string()),
-        }),
-      },
-    },
-    enterpriseSearchRequestHandler.createRequest({
-      path: '/as/engines/:name/source_engines/bulk_create',
-    })
-  );
-  router.delete(
-    {
-      path: '/api/app_search/engines/{name}/source_engines/{source_engine_name}',
-      validate: {
-        params: schema.object({
-          name: schema.string(),
-          source_engine_name: schema.string(),
-        }),
-      },
-    },
-    enterpriseSearchRequestHandler.createRequest({
-      path: '/as/engines/:name/source_engines/:source_engine_name',
-    })
-  );
 }
