+++ x-pack/plugins/enterprise_search/server/routes/app_search/index.ts
@@ -20,6 +20,7 @@ import { registerSchemaRoutes } from './schema';
 import { registerSearchSettingsRoutes } from './search_settings';
 import { registerSearchUIRoutes } from './search_ui';
 import { registerSettingsRoutes } from './settings';
+import { registerSourceEnginesRoutes } from './source_engines';
 import { registerSynonymsRoutes } from './synonyms';
 
 export const registerAppSearchRoutes = (dependencies: RouteDependencies) => {
@@ -30,6 +31,7 @@ export const registerAppSearchRoutes = (dependencies: RouteDependencies) => {
   registerDocumentsRoutes(dependencies);
   registerDocumentRoutes(dependencies);
   registerSchemaRoutes(dependencies);
+  registerSourceEnginesRoutes(dependencies);
   registerCurationsRoutes(dependencies);
   registerSynonymsRoutes(dependencies);
   registerSearchSettingsRoutes(dependencies);
